Disable CGO since it isn't used. (#117)
This change will make the binaries work on any distro, as long as it's linux and the amd64/arm64 architecture matches, removing the need to have a separate binary for each distribution and distro version (no more libc requirement). Disabling CGO does not have an effect on this application functionality. It does however mean that a single amd64-compiled binary will work on every amd64 linux distribution and version.